requirements.txt
Flask
flask-pymongo
flask-jwt-extended
flask-socketio
python-dotenv
cloudinary
authlib
bcrypt
firebase-admin
gunicorn
eventlet
flask-mail